webvh-e2e: move poll_until into helpers, remove polling.py
Phases import poll_until from helpers alongside other harness utilities.

Made-with: Cursor
